web/requirements.txt

pickle-mixin
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
